created unit tests fro the application [Finished #155394163]
