+++ Designs/UI/business.html
@@ -3,7 +3,8 @@
 
 <head>
     <title>BUSINESS REGISTRATION</title>
-    <link rel="stylesheet" type="text/css" href="css/bootstrap.css">
+    <!-- <link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous"> -->
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
 </head>
 
 <body>
@@ -126,8 +127,8 @@
         <h6>WeConnect | Copyright 2018</h6>
     </div>
     <div class=""></div>
-    <script src="jquery-3.3.1.min.js"></script>
-    <script src="js/bootstrap.js"></script>
+    <script src="/static/jquery-3.3.1.min.js"></script>
+    <script src="static/js/bootstrap.js"></script>
 </body>
 
 </html>
